API

API

This page is a dump of all the docstrings found in the code.

MultiSpeciesSpatialEcoModels.calc_species_density — Method

calc_species_density(landscape, numSpecies)

Calculate the density of the species in the landscape matrix

MultiSpeciesSpatialEcoModels.init_multispecies — Method

init_multispecies(fil, col, densidadIni)

Initialize a multispecies model landscape of size fil x col with densities and number of species given by the vector densidadIni and returns a matrix of type Int16.

MultiSpeciesSpatialEcoModels.plot_multispecies — Method

plot_multispecies(landscape,numsp)

plot the landscape matrix of a multispecies model as a heatmap with numsp species

MultiSpeciesSpatialEcoModels.run_birthdeath! — Method

run_birthdeath!(landscape,λ,δ, steps )

Runs the model with landscape grid and parameters vectors λ and δ (birth and death rate) during steps time The model must be initialized previously, it modifies the landscape matrix and returns a matrix of populations over time.

MultiSpeciesSpatialEcoModels.run_birthdeathcol! — Method

run_birthdeathcol!(landscape,λ,δ, steps )

Runs the model with landscape grid and parameters vectors λ δ α (birth death and colonization rate) during steps time The model must be initialized previously, it modifies the landscape matrix and returns a matrix of populations over time.

MultiSpeciesSpatialEcoModels.step_birthdeath! — Method

step_birthdeath!(landscape, λ,δ)

Performs a step of the model a birth-death multispecies model, where landscape is the model grid, and λ and δ are vectors of the birth and death rate parameters. These vectors must have the same length that also defines the number of competing species. The length of λ and δ is the number of species.

The parameters λ and δ could change in any step of the model.

MultiSpeciesSpatialEcoModels.step_birthdeathcol! — Method

step_birthdeathcol!(landscape, λ,δ,α)

Performs a step of the model a birth-death-colonization multispecies model, where landscape is the model grid, and λ and δ are vectors of the birth and death rate parameters. These vectors must have the same length that also defines the number of competing species. The length of λ and δ is the number of species.

The parameters λ and δ could change in any step of the model.